+++ src/main/kotlin/org/patternfly/Title.kt
@@ -3,14 +3,23 @@ package org.patternfly
 import dev.fritz2.dom.html.H
 import dev.fritz2.dom.html.RenderContext
 import kotlinx.coroutines.Job
+import org.patternfly.Level.H1
 import org.w3c.dom.HTMLHeadingElement
 
-// TODO Document me
 // ------------------------------------------------------ dsl
 
+/**
+ * Creates a [Title] component.
+ *
+ * @param level the level of the heading
+ * @param size the size of the heading
+ * @param id the ID of the element
+ * @param baseClass optional CSS class that should be applied to the element
+ * @param content a lambda expression for setting up the component itself
+ */
 public fun RenderContext.title(
-    level: Int = 1,
-    size: Size = Size.XL_2,
+    level: Level = H1,
+    size: Size = level.size,
     id: String? = null,
     baseClass: String? = null,
     content: Title.() -> Unit = {}
@@ -18,15 +27,20 @@ public fun RenderContext.title(
 
 // ------------------------------------------------------ tag
 
+/**
+ * PatternFly [title](https://www.patternfly.org/v4/components/title/design-guidelines/) component.
+ *
+ * A title component applies top and bottom margins, font-weight, font-size, and line-height to titles.
+ */
 public class Title internal constructor(
-    level: Int = 1,
-    size: Size = Size.XL_2,
+    level: Level,
+    size: Size,
     id: String?,
     baseClass: String?,
     job: Job
 ) : PatternFlyComponent<HTMLHeadingElement>,
     H(
-        level,
+        level.level,
         id = id,
         baseClass = classes {
             +ComponentType.Title